+++ eventstore/mongodb/spring/blocking/src/test/java/org/occurrent/eventstore/mongodb/spring/blocking/SpringMongoEventStoreTest.java
@@ -734,7 +734,7 @@ public class SpringMongoEventStoreTest {
 
             @EnabledOnOs(MAC)
             @RepeatedIfExceptionsTest(repeats = 5, suspend = 500)
-            void parallel_writes_to_event_store_throws_WriteConditionNotFulfilledException() {
+            void parallel_writes_to_event_store_throws_WriteConditionNotFulfilledException_when_write_condition_is_not_any() {
                 // Given
                 CyclicBarrier cyclicBarrier = new CyclicBarrier(2);
                 WriteCondition writeCondition = WriteCondition.streamVersionEq(0);
@@ -755,6 +755,46 @@ public class SpringMongoEventStoreTest {
 
                 // Then
                 Awaitility.await().atMost(4, SECONDS).untilAsserted(() -> assertThat(exception).hasValue(new WriteConditionNotFulfilledException("name", 1, writeCondition, "WriteCondition was not fulfilled. Expected version to be equal to 0 but was 1.")));
+            }
+
+            @EnabledOnOs(MAC)
+            @RepeatedIfExceptionsTest(repeats = 5, suspend = 200,  minSuccess = 5)
+            void parallel_writes_to_event_store_does_not_throw_WriteConditionNotFulfilledException_when_write_condition_is_any() {
+                // Given
+                CyclicBarrier cyclicBarrier = new CyclicBarrier(2);
+                WriteCondition writeCondition = WriteCondition.anyStreamVersion();
+                AtomicReference<Throwable> exception = new AtomicReference<>();
+
+                // When
+                new Thread(() -> {
+                    NameWasChanged event = new NameWasChanged(UUID.randomUUID().toString(), now, "Ikk Doe");
+                    try {
+                        await(cyclicBarrier);
+                        persist("name", writeCondition, event);
+                    } catch (Exception e) {
+                        exception.set(e);
+                    }
+                }).start();
+
+                new Thread(() -> {
+                    NameWasChanged event = new NameWasChanged(UUID.randomUUID().toString(), now, "Ikkster Doe");
+                    try {
+                        await(cyclicBarrier);
+                        persist("name", writeCondition, event);
+                    } catch (Exception e) {
+                        exception.set(e);
+                    }
+                }).start();
+
+                // Then
+                Awaitility.await().atMost(4, SECONDS).untilAsserted(() -> {
+                    EventStream<CloudEvent> eventStream = eventStore.read("name");
+                    assertThat(deserialize(eventStream.events()))
+                            .extracting(it -> (NameWasChanged) it)
+                            .extracting(NameWasChanged::getName)
+                            .contains("Ikk Doe", "Ikkster Doe");
+                });
+                assertThat(exception.get()).isNull();
             }
         }
 
@@ -1717,7 +1757,8 @@ public class SpringMongoEventStoreTest {
         return events
                 .map(CloudEvent::getData)
                 // @formatter:off
-                .map(CheckedFunction.unchecked(data -> objectMapper.readValue(data.toBytes(), new TypeReference<Map<String, Object>>() {})))
+                .map(CheckedFunction.unchecked(data -> objectMapper.readValue(data.toBytes(), new TypeReference<Map<String, Object>>() {
+                })))
                 // @formatter:on
                 .map(event -> {
                     Instant instant = Instant.ofEpochMilli((long) event.get("time"));